Async register Vega inspector
