+++ src/platform/plugins/private/vis_types/vega/public/vega_inspector/vega_data_inspector.tsx
@@ -13,7 +13,7 @@ import { EuiTabbedContent, EuiCallOut } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
 import type { InspectorViewProps } from '@kbn/inspector-plugin/public';
 import { css } from '@emotion/react';
-import type { VegaInspectorAdapters } from './vega_inspector';
+import type { VegaInspectorAdapters } from './types';
 import { DataViewer, SignalViewer, SpecViewer } from './components';
 
 export type VegaDataInspectorProps = InspectorViewProps<VegaInspectorAdapters>;
